kanban-tasks-01

+++ backend/real-estate/src/main/java/com/burakcanaksoy/realestate/request/WorkplaceFilterRequest.java
@@ -4,14 +4,16 @@ import com.burakcanaksoy.realestate.model.enums.ListingStatus;
 import com.burakcanaksoy.realestate.model.enums.WorkplaceType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class WorkplaceFilterRequest {
     private String city;
     private String district;
     private String categorySlug;
     private ListingStatus status;
-    private Integer minPrice;
-    private Integer maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private WorkplaceType workplaceType;
     private Integer minSquareMeter;
     private Integer maxSquareMeter;
